Extraterrestrial Exospheres and Surfaces Simulations

For the best UX, please set the documentation theme to light, since all images are optimized for a white background.

Installation

This simulation tool was created as a custom package for the computer language Julia. Please refer to the official guide to install Julia on your machine.

To be able to use the ExESS package, the entire repository has to be downloaded from LRZ-GitLab. Afterwards, the code can be used in a number of different ways. The entire module can be included directly by specifying the path of the downloaded package in the include call.

include(joinpath(PathToPackage, "src", "ExESS.jl"))
using .ExESS

API

ExESS — Module

ExESS

Author Information

Alexander Smolka - alexander.smolka@tum.de