Constants

The following constants are defined in the base of the ExESS package to be used in the calculations of the implemented physical phenomena.

Main.ExESS.AVOGADRO_CONSTANT — Constant

$N_A = 6.02214086 \times 10^{23} \; \text{mol}^{-1}$

Main.ExESS.BOLTZMANN_CONSTANT — Constant

$k_B = 1.38065 \times 10^{-23} \; \text{J K}^{-1}$

Main.ExESS.GRAVITATIONAL_CONSTANT — Constant

$\mu = 6.67408 \times 10^{-11} \; \text{m}^3 \; \text{kg}^{-1} \; \text{s}^{-2}$

Main.ExESS.PLANCK_CONSTANT — Constant

$h = 6.62607004 \times 10^{-34} \; \text{m}^2 \; \text{kg} \; \text{s}^{-1}$

Main.ExESS.UNIVERSAL_GAS_CONSTANT — Constant

$R = 8.3144626181532 \; \text{J} \; \text{K}^{-1} \; \text{mol}^{-1}$

Main.ExESS.ELEMENTARY_CHARGE — Constant

$e = 1.602176634 \times 10^{-19} \; \text{C}$