Coordinate Systems

Currently there are two different custom position and three velocity types implemented, based on a total of three coordinate systems (CS):

fixed, cartesian coordinate system (CCS, subscript "c")

fixed, spherical coordinate system (SCS, subscript "s")

rotating (local) cartesian (spherical) coordinate system (CSCS, subscript "cs")

API

Main.ExESS.CartesianPosition — Type

[1] CartesianPosition(x::Real, y::Real, z::Real)
[2] CartesianPosition(X::Vector{<:Real})
[3] CartesianPosition(xs::SphericalPosition)

Defines a three dimensional, fixed, cartesian position vector. The vector can be created based on positional or vectorial input for x, y, and z, or using a positional vector in another coordinate system like SphericalPosition, which automatically converts the vector into cartesian coordinates.

Fields

x: x position

y: y position

z: z position

Conversions

Let $\vec{x}_c$ be of type CartesianPosition and $\vec{x}_s$ of type SphericalPosition:

\[\vec{x}_c = \begin{bmatrix} x \\ y \\ z \ \end{bmatrix}; \qquad \vec{x}_s = \begin{bmatrix} R \\ \theta \\ \phi \ \end{bmatrix}; \qquad \vec{x}_c \left( \vec{x}_s \right) = \begin{bmatrix} R \cdot \cos\theta \cdot \cos\phi \\ R \cdot \sin\theta \cdot \cos\phi \\ R \cdot \sin\phi \end{bmatrix}\]

Main.ExESS.CartesianSphericalVelocity — Type

[1] CartesianSphericalVelocity(x::Real, y::Real, z::Real)
[2] CartesianSphericalVelocity(V::Vector{<:Real})
[3] CartesianSphericalVelocity(xc::CartesianPosition, vc::CartesianVelocity)
[4] CartesianSphericalVelocity(xs::SphericalPosition, vc::CartesianVelocity)
[5] CartesianSphericalVelocity(xc::CartesianPosition, vs::SphericalVelocity)
[6] CartesianSphericalVelocity(xs::SphericalPosition, vs::SphericalVelocity)

Three dimensional velocity vector in a local cartesian coordinate system along spherical directions. The x-axis points along the radial position, based on the respective body's center. The y-axis points towards positive longitude and the z-axis towards positive latitude.

The velocity vector can be created through direct input for x, y, and z, or

Benchmark Notes

$t_4 \approx 7 \times t_3$

$t_5 \approx 7 \times t_3$

Main.ExESS.CartesianVelocity — Type

[1] CartesianVelocity(x::Real, y::Real, z::Real)
[2] CartesianVelocity(V::Vector{<:Real})
[3] CartesianVelocity(xc::CartesianPosition, vcs::CartesianSphericalVelocity)
[4] CartesianVelocity(xs::SphericalPosition, vcs::SphericalVelocity)
[5] CartesianVelocity(xc::CartesianPosition, vs::CartesianSphericalVelocity)
[6] CartesianVelocity(xs::SphericalPosition, vs::SphericalVelocity)

Main.ExESS.SphericalPosition — Type

[1] SphericalPosition(r::Real, theta::Real, phi::Real)
[2] SphericalPosition(X::Vector{<:Real})
[3] SphericalPosition(xc::CartesianPosition)

Main.ExESS.SphericalVelocity — Type

[1] SphericalVelocity(r::Real, theta::Real, phi::Real)
[2] SphericalVelocity(V::Vector{<:Real})
[3] SphericalVelocity(xc::CartesianPosition, vc::CartesianVelocity)
[4] SphericalVelocity(xs::SphericalPosition, vc::CartesianVelocity)
[5] SphericalVelocity(xc::CartesianPosition, vcs::CartesianSphericalVelocity)
[6] SphericalVelocity(xs::SphericalPosition, vcs::CartesianSphericalVelocity)

Base Extensions

Base.:+ — Function

[1] Base.:+(xc1::CartesianPosition, xc2::CartesianPosition)::CartesianPosition
[2] Base.:+(xc::CartesianPosition, xs::SphericalPosition)::CartesianPosition
[3] Base.:+(xc::CartesianPosition, x::Vector{<:Real})::CartesianPosition
[4] Base.:+(xs1::SphericalPosition, xs2::SphericalPosition)::SphericalPosition
[5] Base.:+(xs::SphericalPosition, xc::CartesianPosition)::SphericalPosition
[6] Base.:+(xs::SphericalPosition, x::Vector{<:Real})::SphericalPosition
[7] Base.:+(vc1::CartesianVelocity, vc2::CartesianVelocity)::CartesianVelocity
[8] Base.:+(vc::CartesianVelocity, v::Vector{<:Real})::CartesianVelocity

Base extension to add simple addition of custom position and velocity vectors. Output type is equal to the first input type with the second variable being converted into the correct format.

Base.:- — Function

[1] Base.:-(xc1::CartesianPosition, xc2::CartesianPosition)::CartesianPosition
[2] Base.:-(xc::CartesianPosition, x::Vector{<:Real})::CartesianPosition
[3] Base.:-(xc::CartesianPosition, x::Vector{<:Real})::CartesianPosition
[4] Base.:-(xs1::SphericalPosition, xs2::SphericalPosition)::SphericalPosition
[5] Base.:-(xs::SphericalPosition, xc::CartesianPosition)::SphericalPosition
[6] Base.:-(xs::SphericalPosition, x::Vector{<:Real})::SphericalPosition
[7] Base.:-(vc1::CartesianVelocity, vc2::CartesianVelocity)::CartesianVelocity
[8] Base.:-(vc::CartesianVelocity, v::Vector{<:Real})::CartesianVelocity

Base extension to add simple subtraction of custom position and velocity vectors. Output type is equal to the first input type with the second variable being converted into the correct format.

Base.:* — Function

[1] Base.:*(xc1::CartesianPosition, xc2::CartesianPosition)::Float32
[2] Base.:*(xc::CartesianPosition, x::Vector{<:Real})::Float32
[3] Base.:*(x::Vector{<:Real}, xc::CartesianPosition)::Float32
[4] Base.:*(xc::CartesianPosition, a::Real)::CartesianPosition
[5] Base.:*(a::Real, xc::CartesianPosition)::CartesianPosition
[6] Base.:*(xs1::SphericalPosition, xs2::SphericalPosition)::Float32
[7] Base.:*(xc::CartesianPosition, xs::SphericalPosition)::Float32
[8] Base.:*(xs::SphericalPosition, xc::CartesianPosition)::Float32
[9] Base.:*(vc1::CartesianVelocity, vc2::CartesianVelocity)::Float32
[10] Base.:*(vc::CartesianVelocity, v::Vector{<:Real})::Float32
[11] Base.:*(v::Vector{<:Real}, vc::CartesianVelocity)::Float32
[12] Base.:*(vc::CartesianVelocity, a::Real)::CartesianVelocity
[13] Base.:*(a::Real, vc::CartesianVelocity)::CartesianVelocity

Base extension to add multiplication of custom position and velocity vectors.

Base.get — Function

[1] Base.get(xc::CartesianPosition)::Vector{Float32}
[2] Base.get(vcs::CartesianSphericalVelocity)::Vector{Float32}
[3] Base.get(vc::CartesianVelocity)::Vector{Float32}
[4] Base.get(xs::SphericalPosition)::Vector{Float32}
[5] Base.get(vs::SphericalVelocity)::Vector{Float32}

Base extension of the get function used to return the fields of the specialized vectors in their usual order $x, y, z$ and $r, \theta, \phi$.

Base.isapprox — Function

[1] Base.isapprox(a::T, b::T; rtol=1e-6) where {T<:Union{Position, Velocity}}
[2] Base.isapprox(xc::CartesianPosition, xs::SphericalPosition; rtol=1e-6)
[3] Base.isapprox(xs::SphericalPosition, xc::CartesianPosition; rtol=1e-6)

Base extension of the isapprox (or ≈) function to compare specialized vectors. The function can handle two inputs of the same type for all special types as well as a comparison between all positional types interchangingly.

Base.isequal — Function

[1] Base.isequal(a::T, b::T; rtol=1e-6) where {T<:Union{Position, Velocity}}
[2] Base.isequal(xc::CartesianPosition, xs::SphericalPosition; rtol=1e-6)
[3] Base.isequal(xs::SphericalPosition, xc::CartesianPosition; rtol=1e-6)

Base extension of the isequal (or ==) function to compare specialized vectors. The function can handle two inputs of the same type for all special types as well as a comparison between all positional types interchangingly.

LinearAlgebra Extensions

Missing docstring.

Missing docstring for dot. Check Documenter's build log for details.

Missing docstring.

Missing docstring for norm. Check Documenter's build log for details.