Trajectory

Manual

Examples

API

Main.ExESS.trajectory — Function

[1] trajectory(x0::CartesianPosition, v0::CartesianVelocity, m::Real, q::Real, ff::Function, terminator::Function, saves=Function[];
    tspan=(0f0,1f10), reltol=1f-4, abstol=1f-4, alg=Tsit5())::ODESolution

Calculates the three-dimensional trajectory of a particle starting at the cartesian position x0, with a cartesian launch velocity of v0, traversing through a forcefield, provided in function ff, based on the mass m and charge q of the particle. The function terminator defines when to stop the differential equations solver, while the functions in the array saves define additional saving points of the solver.

Additional information about ff

Please note that the function ff defines a force field in global cartesian coordinates, as it will be used to calculate the time-dependent acceleration based on Newtonian mechanics.

Additional information about terminator

The terminator function has to be user-defined and needs two inputs, the velocity-position-vector u, and the time t. Note that u is a six-dimensional vector with the first three entries being the velocity and the remaining three entries being the position, both in global cartesian coordinates.

Since terminator is used as a ContinuousCallback (see DifferentialEquations.jl) it has to return 0 at the condition at which the solver shall terminate. One example could be

function terminator(u,t)
    if t <= 0; return 1; end
    if norm(u[4:6]) > LUNAR_RADIUS + 1f7; return 0; end
    return norm(u[4:6]) - LUNAR_RADIUS 
end

Note that, if the condition norm(u[4:6]) > LUNAR_RADIUS is not satisfied, the function returns differentiable output to improve the root-finding algorithms.

Additional information about saves

The last and optional entry saves provides a user-defined number of additional conditions, at which the solver is saving the trajectory's position and time. In order to improve the numerical efficiency, please define the functions with differentiable output (see above).

Specialized functions

trajectory_lunar_gravity(x0::CartesianPosition, v0::CartesianVelocity; reltol=1f-2, alg=Tsit5())::ODESolution

Main.ExESS.trajectory_lunar_gravity — Function

[1] trajectory_lunar_gravity(x0::CartesianPosition, v0::CartesianVelocity; reltol=1f-3)::ODESolution

Specialized version of the trajectory function to apply only for the calculation of trajectories on the Moon under the sole influence of its graviational well.

The solver automatically saves additional timesteps at the transition from Sun to shadow, and vice-versa.

Main.ExESS.gett — Function

[1] gett(traj::ODESolution, idx::Integer)::Float32
[2] gettall(traj::ODESolution)::Vector{Float32}

Returns the time of the ODESolution object, the trajectory traj, at the solution index idx, or at all saved time stamps.

Main.ExESS.getvc — Function

[1] getvc(traj::ODESolution, idx::Integer)::CartesianPosition
[2] getvc(traj::ODESolution)::Vector{CartesianPosition}

Returns the cartesian velocity of the ODESolution object, the trajectory traj, at the solution index idx, or at all saved time stamps.

Main.ExESS.getvs — Function

[1] getvs(traj::ODESolution, idx::Integer)::SphericalVelocity
[2] getvs(traj::ODESolution)::Vector{SphericalVelocity}

Returns the cartesian velocity of the ODESolution object, the trajectory traj, at the solution index idx, or at all saved time stamps.

Main.ExESS.getxc — Function

[1] getxc(traj::ODESolution, idx::Integer)::CartesianPosition
[2] getxc(traj::ODESolution)::Vector{CartesianPosition}

Returns the cartesian position of the ODESolution object, the trajectory traj, at the solution index idx, or at all saved time stamps.

Main.ExESS.getxs — Function

[1] getxs(traj::ODESolution, idx::Integer)::SphericalPosition
[2] getxs(traj::ODESolution)::Vector{SphericalPosition}

Returns the spherical position of the ODESolution object, the trajectory traj, at the solution index idx, or at all saved time stamps.